Datatable Pagination, LengthMenu Düzenlendi ve Görsel İyileştirmeler Yapıldı
İşlemler Butonu Değiştirildi
Kasa Ekranlarındaki Nokta-Virgül Hatası Düzeltildi
Doküman Yüklenmeyen Giderlerin Silinmesi Sırasındaki Hata Düzeltildi
Autocomplete'ler Kapatıldı
Kasa Kayıtlarının Silinmesi Sırasındaki Hata Düzeltildi
Canlı Kopyası DB Eklendi

+++ Controllers/ExpenseController.cs
@@ -504,10 +504,12 @@ namespace ExpenseManagement.Controllers
             await _context.SaveChangesAsync();
 
             var documents = await _context.ExpenseDocuments.FirstOrDefaultAsync(ed => ed.ExpenseId == id);
-            _context.ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
-            _context.ExpenseDocuments.Remove(documents);
-            await _context.SaveChangesAsync();
-
+            if (documents != null)
+            {
+                _context.ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
+                _context.ExpenseDocuments.Remove(documents);
+                await _context.SaveChangesAsync();
+            }
             return Ok(new { Result = true, Message = "Gider Silinmiştir!" });
         }
